+++ dask_jobqueue/lsf.py
@@ -32,10 +32,10 @@ class LSFJob(Job):
         lsf_units=None,
         config_name=None,
         use_stdin=None,
-        **kwargs
+        **base_class_kwargs
     ):
         super().__init__(
-            scheduler=scheduler, name=name, config_name=config_name, **kwargs
+            scheduler=scheduler, name=name, config_name=config_name, **base_class_kwargs
         )
 
         if queue is None: